+++ lib/binary_types.dart
@@ -1,6 +1,7 @@
 library binary_types;
 
 import 'dart:collection';
+import 'dart:io';
 import 'dart:typed_data';
 
 import "package:binary_declarations/binary_declarations.dart";