+++ lib/src/declarations.dart
@@ -18,55 +18,30 @@ class _Declarations {
       throw new ArgumentError.notNull("source");
     }
 
-    var declarations = new Declarations(source, environment: environment);
+    var env = <String, String>{};
+    if (environment != null) {
+      env.addAll(environment);
+    }
+
+    var bitness = types["size_t"].size * 8;
+    env["__OS__"] = Platform.operatingSystem;
+    env["__BITNESS__"] = bitness.toString();
+    var declarations = new Declarations(source, environment: env);
     Declaration declaration;
     try {
       for (declaration in declarations) {
         if (declaration is TypedefDeclaration) {
-          var originalType = declaration.type;
-          var originalBinaryType = _declareType(originalType);
-          for (var synonym in declaration.synonyms) {
-            var binaryType = originalBinaryType;
-            var pointers = synonym.pointers;
-            while (pointers-- > 0) {
-              binaryType = binaryType.ptr();
-            }
-
-            var dimensions = synonym.dimensions;
-            if (dimensions != null) {
-              var length = dimensions.length;
-              for (var i = length - 1; i >= 0; i--) {
-                var dimension = dimensions[i];
-                if (dimension == null) {
-                  dimension = 0;
-                }
-
-                binaryType = binaryType.array(dimension);
-              }
-            }
-
-            var attributes = _Metadata.getAttributes(synonym.metadata, null);
-            attributes = _Metadata.getAttributes(originalType.metadata, attributes);
-            attributes = _Metadata.getAttributes(declaration.metadata, attributes);
-            var align = _Metadata.getAttributeAligned(attributes, null);
-            var packed = _Metadata.getAttributePacked(attributes, false);
-            _defineType(synonym.name, binaryType, align, packed);
-          }
-
+          _declareTypedef(declaration);
         } else if (declaration is StructureDeclaration) {
-          _declareStructure(declaration.type);
+          _declareStructure(declaration);
         } else if (declaration is EnumDeclaration) {
-          _declareEnum(declaration.type);
+          _declareEnum(declaration);
         } else if (declaration is VariableDeclaration) {
-          if (declaration.type is StructureTypeSpecification) {
-            _declareStructure(declaration.type);
-          } else if (declaration.type is EnumTypeSpecification) {
-            _declareEnum(declaration.type);
-          }
+          _declareVariable(declaration);
         }
       }
     } catch (e, s) {
-      BinaryTypeError.declarationError(declaration, e.toString());
+      BinaryTypeError.declarationError(declaration, "$e\n$s");
     }
   }
 
@@ -88,9 +63,149 @@ class _Declarations {
     }
   }
 
-  BinaryType _declareArray(ArrayTypeSpecification type) {
-    var binaryType = _declareType(type.type);
-    for (var dimension in type.dimensions) {
+  void _declareEnum(EnumDeclaration declaration) {
+    var type = declaration.type;
+    var taggedType = type.taggedType;
+    var binaryType = _resolveEnum(type);
+    _registerTaggedType(taggedType, binaryType);
+  }
+
+  StructureMember _declareMember(ParameterDeclaration declaration) {
+    var identifier = declaration.identifier;
+    var type = declaration.type;
+    var binaryType = _resolveType(type);
+    var typeAlign = binaryType.align;
+    var metadata = new _Metadata([identifier.metadata, type.metadata, declaration.metadata]);
+    var align = metadata.aligned;
+    var packed = metadata.packed;
+    if (packed) {
+      if (align == null) {
+        align = 1;
+      }
+    } else {
+      if (align != null) {
+        if (align < typeAlign) {
+          align = typeAlign;
+        }
+      }
+    }
+
+    return new StructureMember(identifier.name, binaryType, align: align, width: declaration.width);
+  }
+
+  void _declareStructure(StructureDeclaration declaration) {
+    var type = declaration.type;
+    var taggedType = type.taggedType;
+    var binaryType = _resolveStructure(type);
+    _registerTaggedType(taggedType, binaryType);
+  }
+
+  BinaryType _declareSynonym(TypeSpecification type, _Out<String> name, BinaryType binaryType) {
+    switch (type.typeKind) {
+      case TypeSpecificationKind.DEFINED:
+        name.value = type.name;
+        break;
+      case TypeSpecificationKind.ARRAY:
+        var arrayType = type as ArrayTypeSpecification;
+        binaryType = _declareSynonym(arrayType.type, name, binaryType);
+        var dimensions = arrayType.dimensions.dimensions;
+        var length = dimensions.length;
+        for (var i = length - 1; i >= 0; i--) {
+          var dimension = dimensions[i];
+          if (dimension == null) {
+            dimension = 0;
+          }
+
+          binaryType = binaryType.array(dimension);
+        }
+
+        break;
+      case TypeSpecificationKind.POINTER:
+        var pointerType = type as PointerTypeSpecification;
+        binaryType = _declareSynonym(pointerType.type, name, binaryType);
+        binaryType = binaryType.ptr();
+        break;
+      default:
+        throw new ArgumentError("Unable declare synonym '${type.typeKind}' $name");
+    }
+
+    return binaryType;
+  }
+
+  void _declareTypedef(TypedefDeclaration declaration) {
+    var type = declaration.type;
+    _registerTypeIfNecessary(type);
+    var originalType = _resolveType(type);
+    for (var synonym in declaration.synonyms) {
+      var name = new _Out<String>();
+      var binaryType = _declareSynonym(synonym, name, originalType);
+      _Metadata metadata;
+      switch (type.typeKind) {
+        case TypeSpecificationKind.DEFINED:
+        case TypeSpecificationKind.FLOAT:
+        case TypeSpecificationKind.INTEGER:
+          // Bug in gcc?
+          metadata = new _Metadata([synonym.metadata, type.metadata]);
+          break;
+        default:
+          metadata = new _Metadata([type.metadata, synonym.metadata]);
+      }
+
+      _defineType(name.value, binaryType, metadata);
+    }
+  }
+
+  void _declareVariable(VariableDeclaration declaration) {
+    _registerTypeIfNecessary(declaration.type);
+  }
+
+  BinaryType _defineType(String synonym, BinaryType binaryType, _Metadata metadata) {
+    var previous = types._types[synonym];
+    if (previous != null && !previous.compatible(binaryType, true)) {
+      BinaryTypeError.unableRedeclareType(synonym);
+    }
+
+    var align = metadata.aligned;
+    var packed = metadata.packed;
+    var copy = binaryType.clone(synonym, align: align, packed: packed);
+    types._types[synonym] = copy;
+    return copy;
+  }
+
+  void _registerTaggedType(TaggedTypeSpecification type, BinaryType binaryType) {
+    var tag = type.tag;
+    _checkTag(binaryType, tag);
+    if (tag != null) {
+      types._tags[tag] = binaryType;
+      types._types[type.name] = binaryType;
+    }
+  }
+
+  void _registerTypeIfNecessary(TypeSpecification type) {
+    BinaryType binaryType;
+    TaggedTypeSpecification taggedType;
+    switch (type.typeKind) {
+      case TypeSpecificationKind.ENUM:
+        taggedType = (type as EnumTypeSpecification).taggedType;
+        binaryType = _resolveEnum(type);
+        break;
+      case TypeSpecificationKind.STRUCTURE:
+        taggedType = (type as StructureTypeSpecification).taggedType;
+        binaryType = _resolveStructure(type);
+        break;
+      default:
+        return;
+    }
+
+    _registerTaggedType(taggedType, binaryType);
+  }
+
+  BinaryType _resolveArray(ArrayTypeSpecification type) {
+    var binaryType = _resolveType(type.type);
+    var dimensions = type.dimensions.dimensions;
+    var length = dimensions.length;
+    for (var i = length - 1; i >= 0; i--) {
+      var dimension = dimensions[i];
       if (dimension == null) {
         dimension = 0;
       }
@@ -101,80 +216,55 @@ class _Declarations {
     return binaryType;
   }
 
-  BinaryType _declareEnum(EnumTypeSpecification type) {
+  BinaryType _resolveEnum(EnumTypeSpecification type) {
     var taggedType = type.taggedType;
-    var attributes = _Metadata.getAttributes(taggedType.metadata, null);
-    attributes = _Metadata.getAttributes(type.metadata, attributes);
-    var align = _Metadata.getAttributeAligned(attributes, null);
+    var metadata = new _Metadata([taggedType.metadata, type.metadata]);
+    var align = metadata.aligned;
+    var packed = metadata.packed;
     var tag = taggedType.tag;
     var values = <String, int>{};
     for (var value in type.values) {
       values[value.name] = value.value;
     }
 
-    var enumType = new EnumType(tag, values, _dataModel, align: align);
-    if (tag != null) {
-      _checkTag(enumType, tag);
-      types._types[enumType.name] = enumType;
-      types._tags[tag] = enumType;
-    }
-
-    return enumType;
+    return new EnumType(tag, values, _dataModel, align: align);
   }
 
-  BinaryType _declareStructure(StructureTypeSpecification type) {
+  BinaryType _resolvePointer(PointerTypeSpecification type) {
+    var binaryType = _resolveType(type.type);
+    return binaryType.ptr();
+  }
+
+  BinaryType _resolveStructure(StructureTypeSpecification type) {
     var taggedType = type.taggedType;
-    var attributes = _Metadata.getAttributes(taggedType.metadata, null);
-    attributes = _Metadata.getAttributes(type.metadata, attributes);
-    var align = _Metadata.getAttributeAligned(attributes, null);
-    var packed = _Metadata.getAttributePacked(attributes, false);
-    var kind = taggedType.kind;
+    var metadata = new _Metadata([taggedType.metadata, type.metadata]);
+    var align = metadata.aligned;
+    var packed = metadata.packed;
+    var tagKind = taggedType.tagKind;
     var tag = taggedType.tag;
     StructureType structureType;
-    if (tag != null) {
-      var name = "$kind $tag";
-      structureType = types._types[name];
-      if (structureType != null) {
-        BinaryTypeError.unableRedeclareType(name);
-      }
+    switch (tagKind) {
+      case TaggedTypeKinds.STRUCT:
+        structureType = new StructType(tag, null, _dataModel, align: align, packed: packed);
+        break;
+      case TaggedTypeKinds.UNION:
+        structureType = new UnionType(tag, null, _dataModel, align: align, packed: packed);
+        break;
     }
 
-    if (structureType == null) {
-      switch (taggedType.kind) {
-        case TaggedTypeKinds.STRUCT:
-          structureType = new StructType(tag, null, _dataModel, align: align, packed: packed);
-          break;
-        case TaggedTypeKinds.UNION:
-          structureType = new UnionType(tag, null, _dataModel, align: align, packed: packed);
-          break;
-      }
-
-      if (tag != null) {
-        _checkTag(structureType, tag);
-        types._types[structureType.name] = structureType;
-        types._tags[tag] = structureType;
-      }
+    BinaryType replacedType;
+    if (tag != null) {
+      replacedType = types._types[taggedType.name];
+      types._types[taggedType.name] = structureType;
     }
 
     var members = <StructureMember>[];
     for (var member in type.members) {
-      var memberType = member.type;
-      var name = member.name;
-      var binaryType = _declareType(memberType);
-      var attributes = _Metadata.getAttributes(memberType.metadata, null);
-      attributes = _Metadata.getAttributes(member.metadata, attributes);
-      var align = _Metadata.getAttributeAligned(attributes, null);
-      // Querying new attributes, since we need ignoring an attribute "packed" in the type
-      attributes = _Metadata.getAttributes(member.metadata, null);
-      var packed = _Metadata.getAttributePacked(attributes, false);
-      if (packed) {
-        if (align == null) {
-          align = 1;
-        }
-      }
+      members.add(_declareMember(member));
+    }
 
-      var newMember = new StructureMember(name, binaryType, align: align, width: member.width);
-      members.add(newMember);
+    if (tag != null) {
+      types._types[taggedType.name] = replacedType;
     }
 
     if (!members.isEmpty) {
@@ -184,40 +274,43 @@ class _Declarations {
     return structureType;
   }
 
-  BinaryType _declareType(TypeSpecification type) {
-    if (type is ArrayTypeSpecification) {
-      return _declareArray(type);
-    } else if (type is DefinedTypeSpecification) {
-      return types._types[type.name];
-    } else if (type is EnumTypeSpecification) {
-      return _declareEnum(type);
-    } else if (type is FloatTypeSpecification) {
-      return types._types[type.name];
-    } else if (type is IntegerTypeSpecification) {
-      return types._types[type.name];
-    } else if (type is PointerTypeSpecification) {
-      return _declareType(type.type).ptr();
-    } else if (type is StructureTypeSpecification) {
-      return _declareStructure(type);
-    } else if (type is TaggedTypeSpecification) {
-      return types._types[type.name];
-    } else if (type is VaListTypeSpecification) {
-      return types._types["..."];
-    } else if (type is VoidTypeSpecification) {
-      return types._types["void"];
-    } else {
-      throw new ArgumentError("Unable declare type by the specification '${type.runtimeType}'");
+  BinaryType _resolveType(TypeSpecification type) {
+    BinaryType binaryType;
+    switch (type.typeKind) {
+      case TypeSpecificationKind.ARRAY:
+        binaryType = _resolveArray(type);
+        break;
+      case TypeSpecificationKind.ENUM:
+        binaryType = _resolveEnum(type);
+        break;
+      case TypeSpecificationKind.DEFINED:
+      case TypeSpecificationKind.FLOAT:
+      case TypeSpecificationKind.INTEGER:
+      case TypeSpecificationKind.TAGGED:
+      case TypeSpecificationKind.VA_LIST:
+      case TypeSpecificationKind.VOID:
+        binaryType = types[type.name];
+        break;
+      case TypeSpecificationKind.POINTER:
+        binaryType = _resolvePointer(type);
+        break;
+      case TypeSpecificationKind.STRUCTURE:
+        binaryType = _resolveStructure(type);
+        break;
+      default:
+        throw new ArgumentError("Unable resolve type '${type.typeKind}'");
     }
+
+    return binaryType;
   }
+}
 
-  BinaryType _defineType(String synonym, BinaryType original, int align, bool packed) {
-    var previous = types._types[synonym];
-    if (previous != null && !previous.compatible(original, true)) {
-      BinaryTypeError.unableRedeclareType(synonym);
-    }
+class _Out<T> {
+  T value;
 
-    var copy = original.clone(synonym, align: align, packed: packed);
-    types._types[synonym] = copy;
-    return copy;
+  _Out([this.value]);
+
+  String toString() {
+    return value.toString();
   }
 }